Fix Travis compiling
