+++ rpcs3/Emu/SysCalls/Modules/sceNp.cpp
@@ -12,22 +12,6 @@
 
 extern Module sceNp;
 
-struct sceNpInternal
-{
-	bool m_bSceNpInitialized;
-	bool m_bScoreInitialized;
-	bool m_bLookupInitialized;
-	bool m_bSceNpUtilBandwidthTestInitialized;
-
-	sceNpInternal()
-		: m_bSceNpInitialized(false),
-		  m_bScoreInitialized(false),
-		  m_bLookupInitialized(false),
-		  m_bSceNpUtilBandwidthTestInitialized(false)
-	{
-	}
-};
-
 sceNpInternal sceNpInstance;
 
 s32 sceNpInit(u32 poolsize, vm::ptr<u32> poolptr)
